requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Pillow == 6.1.0
numpy == 1.17.0
tensorflow == 2.1.0